% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchVessel.R
\name{getCafeTagData}
\alias{getCafeTagData}
\title{Get White Shark Tag data from Cafe'}
\usage{
getCafeTagData(date1, date2)
}
\arguments{
\item{date1}{initial date}

\item{date2}{final date}
}
\description{
Get White Shark Tag data from Cafe'
}
