% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpeed.R
\name{getClosestPosition}
\alias{getClosestPosition}
\title{Get Closest Positions between Fish and Ships}
\usage{
getClosestPosition(out)
}
\arguments{
\item{out}{output object from \code{searchVessel} function}
}
\description{
Get Closest Positions between Fish and Ships
}
