% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrift.R
\name{getDrift}
\alias{getDrift}
\title{Get Drifting Positions of PAT tags}
\usage{
getDrift(con, ptt = 173165)
}
\arguments{
\item{con}{a connection to the OIB - TOPP database.}

\item{ptt}{is the argos ID. If the tag was reused that it is possible to have multiple tags for the sam ptt. In that case use latest date.}
}
\description{
This function connects to the TOPP database and extract drifting positions of floating PATs based on their ptt.
}
