% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpeed.R
\name{getSpeed}
\alias{getSpeed}
\title{Get Speed from Satellite Track}
\usage{
getSpeed(dat, track = "PAT")
}
\arguments{
\item{dat}{a data.frame with a timeseries of location points from PAT tags or AIS track. It has to have coordinates (lat and lon) and timestamps (datetime) in GMT}

\item{track}{whether it is a PAT track or AIS}
}
\description{
Get Speed from Satellite Track
}
