% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchVessel.R
\name{getVesselTracksBox}
\alias{getVesselTracksBox}
\title{Get Fishing Vessel Tracks in in Bounding Box}
\usage{
getVesselTracksBox(con, project, bbox = c(18, 40, -145, -120),
  date1 = "2018-04-25", date2 = "2018-05-06",
  gears = c("drifting_longlines", "trawlers", "purse_seines",
  "unknown_fishing", "squid_jigger", "pole_and_line", "drifting_gillnets",
  "no_idea_what_it_is", "research other_fishing", "set_longlines", "troller",
  "set_gillnets", "pots_and_traps", "tuna_purse_seines"))
}
\arguments{
\item{con}{a connection to the OIB - TOPP database.}

\item{project}{Google Cloud Platform project name. This is needed to query the GFW database in bigQuery.}

\item{gears}{fishing gears to search}

\item{time1}{initial date of observation period: "YYY-mm-dd"}

\item{time2}{initial date of observation period: "YYY-mm-dd"}
}
\description{
This function searches for vessels nearby popup events. It uses the Global Fishing Watch (GFW) AIS database and to Tag of Pelagic Predator (TOPP, `http://gtopp.org/`) database and searches within a time window specified by the user, centered on the popup date, with a spatial box specified by the user in latitude and longitude degrees.
}
