% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locationErr}
\alias{locationErr}
\title{Location Error.}
\format{A data frame with 6 rows and 3 variables:
\describe{
  \item{Location.class}{}
  \item{Location.type}{Argos or GPS}
  \item{Location.error}{in meters}
}}
\usage{
locationErr
}
\description{
A dataset containing the location errors in meters corresponding to different location classes.
}
\keyword{datasets}
