% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchVessel.R
\name{mapInteraction}
\alias{mapInteraction}
\title{Map Possible Tag - Vessel Interaction}
\usage{
mapInteraction(out)
}
\arguments{
\item{out}{output of \code{searchVessel()}}
}
\description{
This function uses the output of searchVessel and map the locations of te tags and vessel tracks on a leaflet map.
}
