% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpeed.R
\name{plotSpeed}
\alias{plotSpeed}
\title{Plot Speed from Satellite Track and Vessels}
\usage{
plotSpeed(out)
}
\arguments{
\item{out}{output object from searchVessel}
}
\description{
Plot Speed from Satellite Track and Vessels
}
