% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchVessel.R
\name{searchVessel}
\alias{searchVessel}
\title{Search for Vessels Close to Popups}
\usage{
searchVessel(con, ptt = "153499", spatWin = 1, tempWin = 24,
  Plot = FALSE, gears = c("drifting_longlines", "trawlers"))
}
\arguments{
\item{con}{a connection to the TOPP database}

\item{ptt}{is the argos ID. If the tag was reused that it is possible to have multiple tags for the sam ptt. In that case use latest date.}

\item{spatWin}{spatial window in degrees}

\item{tempWin}{temporal window in hours}

\item{gears}{fishing gears to search}

\item{rtagnumber}{tag number code in the TOPP database. This is the serial number from the wildlife computers}
}
\description{
search for nearby vessels closby popup events. It searches within a time window of three days centered on popupdate and spatial box of one degree in lat and lon.
}
