% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchVessel.R
\name{searchVesselInBbox}
\alias{searchVesselInBbox}
\title{Search for Vessels in bbox}
\usage{
searchVesselInBbox(con, project, bbox = c(0, 70, -180, -95), year = "2014",
  month1 = "01", day1 = "01", year2 = year, month2 = "12",
  day2 = "31", gears = c("drifting_longlines", "trawlers", "purse_seines",
  "unknown_fishing", "squid_jigger", "pole_and_line", "drifting_gillnets",
  "no_idea_what_it_is", "research other_fishing", "set_longlines", "troller",
  "set_gillnets", "pots_and_traps"))
}
\arguments{
\item{con}{a connection to the OIB - TOPP database.}

\item{project}{Google Cloud Platform project name. This is needed to query the GFW database in bigQuery.}

\item{gears}{fishing gears to search}

\item{ptt}{is the argos ID. If the tag was reused that it is possible to have multiple tags for the sam ptt. In that case use latest date.}

\item{spatWin}{spatial window in degrees}

\item{tempWin}{temporal window in hours}
}
\description{
This function searches for vessels nearby popup events. It uses the Global Fishing Watch (GFW) AIS database and to Tag of Pelagic Predator (TOPP, `http://gtopp.org/`) database and searches within a time window specified by the user, centered on the popup date, with a spatial box specified by the user in latitude and longitude degrees.
}
