% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapInteraction.R
\name{mapInteraction}
\alias{mapInteraction}
\title{Map Possible Tag - Vessel Interaction}
\usage{
mapInteraction(out, ClosestDist = TRUE)
}
\arguments{
\item{out}{output of \code{searchVessel()}}

\item{CosestDis}{indicate closest point in time and space}
}
\description{
This function uses the output of searchVessel and map the locations of te tags and vessel tracks on a leaflet map.
}
